Update MetricsConstants and MetricsHook to use long instead of double

+++ src/OpenFeature.Contrib.Hooks.Otel/MetricsConstants.cs
@@ -2,7 +2,7 @@ namespace OpenFeature.Contrib.Hooks.Otel
 {
     internal static class MetricsConstants
     {
-        internal const string Unit = "double";
+        internal const string Unit = "long";
 
         internal const string ActiveCountName = "feature_flag.evaluation_active_count";
         internal const string RequestsTotalName = "feature_flag.evaluation_requests_total";
